fix: database column type routeId was not big enough to handle the new routes

+++ dbmodel.sql
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS `destination` (
 
 --normal direction is from->to, reverse is to->from
 CREATE TABLE IF NOT EXISTS `claimed_routes` (
-  `route_id` TINYINT unsigned NOT NULL,
+  `route_id` smallint unsigned NOT NULL,
   `player_id` int(11) NOT NULL,
   `reverse_direction` TINYINT unsigned NOT NULL DEFAULT FALSE, 
   PRIMARY KEY (`route_id`, `player_id`)
